requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
httpx>=0.26.0
openai>=1.12.0
anthropic>=0.18.0
pydantic>=2.6.0
pyyaml>=6.0.1
peewee>=3.17.0
aiosqlite>=0.19.0
rich>=13.7.0
python-dotenv>=1.0.1
mcp>=1.0.0
pycryptodome>=3.20.0
yfinance>=0.2.36

pypdf>=4.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0
defusedxml>=0.7.0
python-docx>=1.1.0
python-pptx>=0.6.23
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
